+++ CopilotKit/examples/next-openai/src/app/api/copilotkit/route.ts
@@ -42,6 +42,7 @@ const runtime = new CopilotRuntime({
       },
     },
   ],
+  remoteActions: [{ url: "http://localhost:8000/copilotkit" }],
 });
 
 export const POST = async (req: NextRequest) => {