+++ CopilotKit/packages/runtime-client-gql/src/graphql/definitions/mutations.ts
@@ -50,6 +50,11 @@ export const generateCopilotResponseMutation = graphql(/** GraphQL **/ `
           actionExecutionId
           actionName
         }
+        ... on AgentStateMessageOutput {
+          threadId
+          state
+          running
+        }
       }
     }
   }