+++ CopilotKit/packages/runtime/src/lib/runtime/copilot-runtime.ts
@@ -13,23 +13,25 @@
  */
 
 import { Action, actionParametersToJsonSchema, Parameter } from "@copilotkit/shared";
-import { RemoteChain, RemoteChainParameters, CopilotServiceAdapter } from "../service-adapters";
-import { MessageInput } from "../graphql/inputs/message.input";
-import { ActionInput } from "../graphql/inputs/action.input";
-import { RuntimeEventSource } from "../service-adapters/events";
-import { convertGqlInputToMessages } from "../service-adapters/conversion";
-import { Message } from "../graphql/types/converted";
-import { ForwardedParametersInput } from "../graphql/inputs/forwarded-parameters.input";
+import { RemoteChain, RemoteChainParameters, CopilotServiceAdapter } from "../../service-adapters";
+import { MessageInput } from "../../graphql/inputs/message.input";
+import { ActionInput } from "../../graphql/inputs/action.input";
+import { RuntimeEventSource } from "../../service-adapters/events";
+import { convertGqlInputToMessages } from "../../service-adapters/conversion";
+import { Message } from "../../graphql/types/converted";
+import { ForwardedParametersInput } from "../../graphql/inputs/forwarded-parameters.input";
+import { setupRemoteActions, RemoteActionDefinition } from "./remote-actions";
+import { GraphQLContext } from "../integrations/shared";
 
 interface CopilotRuntimeRequest {
   serviceAdapter: CopilotServiceAdapter;
   messages: MessageInput[];
   actions: ActionInput[];
   outputMessagesPromise: Promise<Message[]>;
-  properties: any;
   threadId?: string;
   runId?: string;
   publicApiKey?: string;
+  graphqlContext: GraphQLContext;
   forwardedParameters?: ForwardedParametersInput;
 }
 
@@ -106,6 +108,11 @@ export interface CopilotRuntimeConstructorParams<T extends Parameter[] | [] = []
   actions?: ActionsConfiguration<T>;
 
   /*
+   * A list of remote actions that can be executed.
+   */
+  remoteActions?: RemoteActionDefinition[];
+
+  /*
    * An array of LangServer URLs.
    */
   langserve?: RemoteChainParameters[];
@@ -113,6 +120,7 @@ export interface CopilotRuntimeConstructorParams<T extends Parameter[] | [] = []
 
 export class CopilotRuntime<const T extends Parameter[] | [] = []> {
   public actions: ActionsConfiguration<T>;
+  private remoteActionDefinitions: RemoteActionDefinition[];
   private langserve: Promise<Action<any>>[] = [];
   private onBeforeRequest?: OnBeforeRequestHandler;
   private onAfterRequest?: OnAfterRequestHandler;
@@ -125,21 +133,110 @@ export class CopilotRuntime<const T extends Parameter[] | [] = []> {
       this.langserve.push(remoteChain.toAction());
     }
 
+    this.remoteActionDefinitions = params?.remoteActions || [];
+
     this.onBeforeRequest = params?.middleware?.onBeforeRequest;
     this.onAfterRequest = params?.middleware?.onAfterRequest;
   }
 
+  // async processAgentRequest(request: CopilotRuntimeRequest): Promise<CopilotRuntimeResponse> {
+  //   const { messages, outputMessagesPromise, graphqlContext } = request;
+
+  //   const message = request.messages.slice(-1)[0].agentMessage!;
+
+  //   const agentName = message.agentName;
+  //   const threadId = message.threadId;
+  //   const state = JSON.parse(message.state);
+  //   const nodeName = message.nodeName;
+
+  //   // Fetch remote actions
+  //   const remoteActions = await fetchRemoteActionLocations({
+  //     remoteActionDefinitions: this.remoteActionDefinitions,
+  //     graphqlContext,
+  //   });
+
+  //   const url = remoteActions.get(agentName);
+  //   if (!url) {
+  //     throw new Error(`Action location for agent name ${agentName} not found.`);
+  //   }
+
+  //   const inputMessages = convertGqlInputToMessages(messages);
+
+  //   await this.onBeforeRequest?.({
+  //     threadId,
+  //     runId: undefined,
+  //     inputMessages,
+  //     properties: graphqlContext.properties,
+  //   });
+  //   try {
+  //     const eventSource = new RuntimeEventSource();
+
+  //     const result = await executeAgent({
+  //       agentName,
+  //       nodeName,
+  //       threadId,
+  //       state,
+  //       url,
+  //       graphqlContext,
+  //       logger: graphqlContext.logger,
+  //     });
+
+  //     eventSource.stream(async (eventStream$) => {
+  //       eventStream$.sendAgentMessage(
+  //         result.threadId,
+  //         agentName,
+  //         result.nodeName,
+  //         result.state,
+  //         result.running,
+  //       );
+  //       eventStream$.complete();
+  //     });
+
+  //     outputMessagesPromise
+  //       .then((outputMessages) => {
+  //         this.onAfterRequest?.({
+  //           threadId: result.threadId,
+  //           runId: undefined,
+  //           inputMessages,
+  //           outputMessages,
+  //           properties: graphqlContext.properties,
+  //         });
+  //       })
+  //       .catch((_error) => {});
+
+  //     return {
+  //       threadId: result.threadId,
+  //       runId: undefined,
+  //       eventSource,
+  //       actions: [],
+  //     };
+  //   } catch (error) {
+  //     console.error("Error getting response:", error);
+  //     throw error;
+  //   }
+  // }
+
   async process(request: CopilotRuntimeRequest): Promise<CopilotRuntimeResponse> {
+    // if (request.messages.length > 0) {
+    //   const [lastMessage] = request.messages.slice(-1);
+    //   if (lastMessage.agentMessage) {
+    //     return this.processAgentRequest(request);
+    //   }
+    // }
+
     const {
       serviceAdapter,
-      messages,
+      messages: rawMessages,
       actions: clientSideActionsInput,
       threadId,
       runId,
-      properties,
       outputMessagesPromise,
+      graphqlContext,
       forwardedParameters,
     } = request;
+
+    const messages = rawMessages.filter((message) => !message.agentStateMessage);
+    const inputMessages = convertGqlInputToMessages(messages);
     const langserveFunctions: Action<any>[] = [];
 
     for (const chainPromise of this.langserve) {
@@ -151,10 +248,19 @@ export class CopilotRuntime<const T extends Parameter[] | [] = []> {
       }
     }
 
-    const configuredActions =
-      typeof this.actions === "function" ? this.actions({ properties }) : this.actions;
+    // Fetch remote actions
+    const remoteActions = await setupRemoteActions({
+      remoteActionDefinitions: this.remoteActionDefinitions,
+      graphqlContext,
+      messages: inputMessages,
+    });
 
-    const actions = [...configuredActions, ...langserveFunctions];
+    const configuredActions =
+      typeof this.actions === "function"
+        ? this.actions({ properties: graphqlContext.properties })
+        : this.actions;
+
+    const actions = [...configuredActions, ...langserveFunctions, ...remoteActions];
 
     const serverSideActionsInput: ActionInput[] = actions.map((action) => ({
       name: action.name,
@@ -166,13 +272,12 @@ export class CopilotRuntime<const T extends Parameter[] | [] = []> {
       ...serverSideActionsInput,
       ...clientSideActionsInput,
     ]);
-    const inputMessages = convertGqlInputToMessages(messages);
 
     await this.onBeforeRequest?.({
       threadId,
       runId,
       inputMessages,
-      properties,
+      properties: graphqlContext.properties,
     });
 
     try {
@@ -194,7 +299,7 @@ export class CopilotRuntime<const T extends Parameter[] | [] = []> {
             runId: result.runId,
             inputMessages,
             outputMessages,
-            properties,
+            properties: graphqlContext.properties,
           });
         })
         .catch((_error) => {});